refactor!: remove unwanted functions from public api
